Moved ScatterPlot to /charts/
